Backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.27.0
motor>=3.3.2
pydantic>=2.6.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
typing-extensions==4.8.0
